Check Constraints, Alternate Email, and Inventory Table Additions
Added check constraints to be sure that Movie runTime is >0 and
Inventory quantityAtKiosk is >= 0.  Added an optional alternateEmail
field to the Customer table.  Added quantityAtKiosk and inStock fields
for the Inventory table to keep track of how many copies of each movie
are at each kiosk.  A SQL command to update the Inventory.inStock value
based on quantityAtKiosk was also added.