New ListAllMovies Form
Added a new form which will List All Movies in a list box.  Currently
only movie titles are listed.  The form is currently accessible by (1)
starting the program, going to the Movies form, and then pressing
CTRL+ALT+L or (2) starting the program, going to the Movies form, and
pressing the "List All Movies" button at the top.  Also added a "Search
By Movie Title" button on the ListAllMovies form to link back to the
regular Movies form.