Added DeleteACustomer Form - Other Minor Changes
Added a form for deleting a customer from the database.  The form has a
menuStrip which links to all other active pages.  The comboBox for
customer names is set to fill when the form is loaded, and "DELETE
CUSTOMER" button functionality will be added later.  Made a few other
minor changes to the appearance of some forms, including adding buttons
to the top of the "ListAllCustomers" form.